Refactor `TripRequestDto` and mapping configurations
- Remove `DriverName` from `TripRequestDto`.
- Add `TripRequestForVehicles` DTO.
- Update `MappingProfile` to map `DriverName` for `TripRequestForVehicles` instead of `TripRequestDto`.
