+++ src/VMTS.API/Dtos/Trip/TripRequestForVehicles.cs
@@ -1,13 +1,14 @@
-﻿using VMTS.Core.Entities.Trip;
+using VMTS.Core.Entities.Trip;
 
-namespace VMTS.API.Dtos;
+namespace VMTS.API.Dtos.Trip;
 
-public class TripRequestDto
+public class TripRequestForVehicles
 {
     public string DriverId { get; set; }
 
     public string VehicleId { get; set; }
 
+    public string DriverName { get; set; }
     public TripType TripType { get; set; }
 
     public DateTime Date { get; set; }
